tdf#161837 use double instead of Fraction class
replaced use of Fraction in SfxInPlaceClient::SetObjAreaAndScale and
related methods SetSizeScale, GetScaleWidth, GetScaleHeight;
replaced attributes m_aScaleWidth and m_aScaleHeight from
SfxInPlaceClient_Impl with m_fScaleWidth and m_fScaleHeight

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/201310
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/202809
Tested-by: Tomaž Vajngerl <quikee@gmail.com>

+++ svx/source/svdraw/svdoole2.cxx
@@ -1600,7 +1600,7 @@ void SdrOle2Obj::ImpSetVisAreaSize()
                 if ( pClient )
                 {
                     tools::Rectangle aScaleRect(getRectangle().TopLeft(), aObjAreaSize);
-                    pClient->SetObjAreaAndScale( aScaleRect, aScaleWidth, aScaleHeight);
+                    pClient->SetObjAreaAndScale( aScaleRect, double(aScaleWidth), double(aScaleHeight));
                 }
                 else
                 {
